fastapi/frontend/requirements.txt
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
streamlit>=1.38,<2.0